+++ src/flatlandian/int_vector2.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from dataclasses import dataclass, fields
 from typing import TYPE_CHECKING
 
-from pygame import Vector2
+from pygame.math import Vector2
 
 if TYPE_CHECKING:
     from collections.abc import Sized
@@ -25,27 +25,39 @@ def _ensure_2_elements(value: Sized) -> bool:
 class IntVector2:
     """A 2-dimensional integer vector.
 
-    `IntVector2() -> IntVector2(0, 0)`
+    Supports `len` (always 2) and indexing.
+    Conforms to `pygame.typing.IntPoint`.
 
-    `IntVector2(x: int, y: int) -> IntVector2`
+    Constructors
+    ------------
+    ```
+    IntVector2() -> IntVector2(0, 0)
+    IntVector2(int, int) -> IntVector2
+    IntVector2(x: int, y: int) -> IntVector2
+    ```
 
-    `IntVector2(int, int) -> IntVector2`
+    Attempting to construct with `float` values raises `TypeError`.
 
-    Attempting to construct with `float` parameters raises `TypeError`.
+    Equivalents to these `pygame.math.Vector2` constructors aren't yet supported:
+    ```
+    Vector2(int) -> Vector2
+    Vector2(Vector2) -> Vector2
+    Vector2((x, y)) -> Vector2
+    ```
+    Arithmetic
+    ----------
+    Supports element-wise addition and subtraction of another
+    `IntVector2` or `pygame.typing.IntPoint` i.e. an `int` pair.
+    Returns `IntVector2`.
 
-    Equivalents to these `pygame.Vector2` constructors aren't yet supported:
-
-    `Vector2(int) -> Vector2`
-
-    `Vector2(Vector2) -> Vector2`
-
-    `Vector2((x, y)) -> Vector2`
+    Supports scalar multiplication and floor division by `int`.
+    Returns `IntVector2`.
     """
 
     x: int = 0
-    """x coordinate."""
+    """x coordinate. Also available via index 0 or -1."""
     y: int = 0
-    """y coordinate."""
+    """y coordinate. Also available via index 1 or -2."""
 
     def __post_init__(self) -> None:
         if not isinstance(self.x, int) or not isinstance(self.y, int):
@@ -54,13 +66,13 @@ class IntVector2:
 
     @classmethod
     def from_point(cls, value: IntPoint) -> IntVector2:
-        """Construct `IntVector2` from `IntPoint` i.e. pair of `int`s."""
+        """Construct `IntVector2` from `pygame.typing.IntPoint`, i.e. an `int` pair."""
         _ensure_2_elements(value)
         return cls(value[0], value[1])
 
     @property
     def as_vector2(self) -> Vector2:
-        """Return `pygame.Vector2`, i.e. float coordinates.
+        """Return `pygame.math.Vector2`, i.e. float coordinates.
 
         For compatibility with Pygame functions.
         """