+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project attempts to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED - TBC
+## [0.2.0] - 2025-01-22
 
 ### Added
 
@@ -59,7 +59,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 Initial release
 
-[UNRELEASED]: https://github.com/recreational-projects/flatlandian/compare/v0.1.3...HEAD
+[UNRELEASED]: https://github.com/recreational-projects/flatlandian/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/recreational-projects/flatlandian/compare/v0.1.4...v0.2.0
 [0.1.4]: https://github.com/recreational-projects/flatlandian/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/recreational-projects/flatlandian/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/recreational-projects/flatlandian/compare/v0.1.1...v0.1.2